feat(appeals): rename linked appeals utils file (a2-3075 - part 1)
