+++ appeals/web/src/server/appeals/appeal-details/accordions/common/case-documentation.js
@@ -1,4 +1,4 @@
-import { isChildAppeal } from '#lib/mappers/utils/is-child-appeal.js';
+import { isChildAppeal } from '#lib/mappers/utils/is-linked-appeal.js';
 import { isDefined } from '#lib/ts-utilities.js';
 
 /**
