+++ appeals/web/src/server/lib/mappers/data/appeal/submappers/costs-appellant-decision.mapper.js
@@ -4,7 +4,7 @@ import { isStatePassed } from '#lib/appeal-status.js';
 import { permissionNames } from '#environment/permissions.js';
 import { addBackLinkQueryToUrl } from '#lib/url-utilities.js';
 import { baseUrl } from '#appeals/appeal-details/issue-decision/issue-decision.utils.js';
-import { isChildAppeal } from '#lib/mappers/utils/is-child-appeal.js';
+import { isChildAppeal } from '#lib/mappers/utils/is-linked-appeal.js';
 
 /** @type {import('../mapper.js').SubMapper} */
 export const mapCostsAppellantDecision = ({ appealDetails, currentRoute, session, request }) => {
