+++ appeals/web/src/server/lib/mappers/data/appeal/submappers/lpa-questionnaire-due-date.mapper.js
@@ -6,7 +6,7 @@ import {
 } from '@pins/appeals/constants/support.js';
 import { APPEAL_CASE_STATUS } from '@planning-inspectorate/data-model';
 import { APPEAL_TYPE } from '@pins/appeals/constants/common.js';
-import { isChildAppeal } from '#lib/mappers/utils/is-child-appeal.js';
+import { isChildAppeal } from '#lib/mappers/utils/is-linked-appeal.js';
 
 /** @type {import('../mapper.js').SubMapper} */
 export const mapLpaQuestionnaireDueDate = ({
