--- appeals/web/src/server/lib/mappers/utils/is-child-appeal.js
@@ -1,10 +1,0 @@
-import config from '#environment/config.js';
-
-/**
- *
- * @param {*} appeal
- * @returns {boolean}
- */
-export function isChildAppeal(appeal) {
-	return appeal.isChildAppeal && config.featureFlags.featureFlagLinkedAppeals;
-}
