+++ appeals/web/src/server/lib/mappers/utils/is-linked-appeal.js
@@ -1,0 +1,28 @@
+import config from '#environment/config.js';
+
+/**
+ *
+ * @param {*} appeal
+ * @returns {boolean}
+ */
+export function isChildAppeal(appeal) {
+	return appeal.isChildAppeal && config.featureFlags.featureFlagLinkedAppeals;
+}
+
+/**
+ *
+ * @param {*} appeal
+ * @returns {boolean}
+ */
+export function isParentAppeal(appeal) {
+	return appeal.isParentAppeal && config.featureFlags.featureFlagLinkedAppeals;
+}
+
+/**
+ *
+ * @param {*} appeal
+ * @returns {boolean}
+ */
+export default function isLinkedAppeal(appeal) {
+	return isChildAppeal(appeal) || isParentAppeal(appeal);
+}
