+++ appeals/web/src/server/lib/mappers/utils/map-status-dependent-notifications.js
@@ -6,7 +6,7 @@ import { getRequiredActionsForAppeal } from './required-actions.js';
 import { addBackLinkQueryToUrl } from '#lib/url-utilities.js';
 import { generateIssueDecisionUrl } from '#appeals/appeal-details/issue-decision/issue-decision.utils.js';
 import config from '../../../../../environment/config.js';
-import { isChildAppeal } from '#lib/mappers/utils/is-child-appeal.js';
+import { isChildAppeal } from '#lib/mappers/utils/is-linked-appeal.js';
 
 /** @typedef {import('./required-actions.js').AppealRequiredAction} AppealRequiredAction */
 /** @typedef {import('../components/index.js').NotificationBannerDefinitionKey} NotificationBannerDefinitionKey */
